data_extraction/dependency_analysis/requirements.txt
# General dependencies
pandas
numpy
matplotlib
seaborn
scipy
statsmodels
requests
json

# Git operations
GitPython

# Argument parsing
argparse

# JSON and CSV processing
csv
